+++ controllers/messagesController.js
@@ -9,6 +9,15 @@ class MessagesController {
         if(!message || !outgoing_id || !incoming_id) {
             return next(ApiError.badRequest('Не задано одно из обязательных полей'));
         }
+
+        if(message.length > 1000 || message.length <= 0) {
+            return next(ApiError.badRequest('Слишком большое сообщение'));
+        }
+
+        if(req.decodedToken.email !== outgoing_id) {
+            return next(ApiError.badRequest('Ты чего тут удумал еблоид?'));
+        }
+
         const msg = await Messages.create({message, outgoing_id, incoming_id, viewed: false});
         return res.json(msg);
     }
@@ -17,6 +26,11 @@ class MessagesController {
         if(!firstUserId || !secondUserId) {
             return next(ApiError.badRequest('Не задано одно из обязательных полей'));
         }
+
+        if(req.decodedToken.email !== firstUserId && req.decodedToken.email !== secondUserId) {
+            return next(ApiError.badRequest('Ты чего тут удумал еблоид?'));
+        }
+
         page = page || 1;
         limit = limit || 15;
         let offset = page * limit - limit;
@@ -43,6 +57,11 @@ class MessagesController {
         if(!queryParameter || !id) {
             return next(ApiError.badRequest('Не задано одно из обязательных полей'));
         }
+        
+        if(req.decodedToken.email !== id) {
+            return next(ApiError.badRequest('Ты чего тут удумал еблоид?'));
+        }
+
         page = page || 1;
         limit = limit || 20;
         let offset = page * limit - limit;
@@ -54,6 +73,16 @@ class MessagesController {
         if(!message || !id) {
             return next(ApiError.badRequest('Не задано одно из обязательных полей'));
         }
+
+        if(message.length > 1000 || message.length <= 0) {
+            return next(ApiError.badRequest('Слишком большое сообщение'));
+        }
+
+        const checkMessage = await Messages.findOne({where: {id: id}});
+        if(checkMessage.outgoing_id !== req.decodedToken.email) {
+            return next(ApiError.badRequest('Ты чего тут удумал еблоид?'));
+        }
+
         const updatedMessage = await Messages.update({message}, {where: {id: id}})
         return res.json(updatedMessage)
     }
@@ -70,6 +99,12 @@ class MessagesController {
         if(!id) {
             return next(ApiError.badRequest('Не задано обязательное поле'));
         }
+        
+        const checkMessage = await Messages.findOne({where: {id: id}});
+        if(checkMessage.outgoing_id !== req.decodedToken.email) {
+            return next(ApiError.badRequest('Ты чего тут удумал еблоид?'));
+        }
+
         const deletedMessage = await Messages.destroy({where: {id: id}})
         return res.json(deletedMessage)
     }
@@ -78,6 +113,11 @@ class MessagesController {
         if(!firstUserId || !secondUserId) {
             return next(ApiError.badRequest('Не задано одно из обязательных полей'));
         }
+
+        if(firstUserId !== req.decodedToken.email && secondUserId !== req.decodedToken.email) {
+            return next(ApiError.badRequest('Ты чего тут удумал еблоид?'));
+        }
+
         const messagesCount = await Messages.count({
         where: {
             [Op.or]: [ 
@@ -101,6 +141,11 @@ class MessagesController {
         if(!user_id) {
             return next(ApiError.badRequest('Не задано обязательное поле'));
         }
+
+        if(user_id !== req.decodedToken.email) {
+            return next(ApiError.badRequest('Ты чего тут удумал еблоид?'));
+        }
+
         const messagesCount = await Messages.count({where: {[Op.or]: [{outgoing_id: user_id}, {incoming_id: user_id}]}});
         return res.json(messagesCount);
     }
@@ -109,6 +154,11 @@ class MessagesController {
         if(!user_id) {
             return next(ApiError.badRequest('Не задано обязательное поле'));
         }
+        
+        if(user_id !== req.decodedToken.email) {
+            return next(ApiError.badRequest('Ты чего тут удумал еблоид?'));
+        }
+
         const message = await Messages.findOne({where: {incoming_id: user_id, viewed: false}})
         if(message) return res.json(true);
         else return res.json(false);
